Desktop Video 11.1 SDK (01 Apr 2019) This SDK provides developer support for Desktop Video 11.1 that allows updating of hardware control and software interfaces for Desktop Video products.
Desktop Video 11.1 (01 Apr 2019)
This update adds support for DeckLink Quad HDMI Recorder and also enables multi-channel configuration support for DeckLink 8K Pro from within Setup Utility. Desktop Video 11.1 update also adds HDR support to DeckLink 4K Extreme 12G via SDI. This includes Rec. 2020, HDR metadata packing (ST2108-1) on SDI, and HLG and PQ transfer characteristics on both SDI (ST2082-10) and HDMI (CEA 861-G).

+++ Win/Samples/PlaybackStills/ImageLoaderWin.cpp
@@ -166,7 +166,7 @@ HRESULT ImageLoader::ConvertPNGToDeckLinkVideoFrame(const std::string& pngFilena
 			goto bail;
 	}
 
-	rectLock = { 0, 0, imageWidth, imageHeight };
+	rectLock = { 0, 0, static_cast<INT>(imageWidth), static_cast<INT>(imageHeight) };
 
 	// Create bitmap lock so we can access the pixel data directly.
 	result = bitmapFrame->Lock(&rectLock, WICBitmapLockRead, &bitmapLock);
